Brains Data Import Function
This is a script for a function that is imported into other analyses scripts. This function is written to process data for multiple brains from LifeCanvas Technologies, our data vendor that processes the brain clarification and light sheet imaging for our lab. 
The function works by having all data files intended for analysis in a single directory. With the only argument of the function being the directory in which the files are stored, the function takes the first and second half of the dataframe, being the right and left hemispheres, respectively, and combines them together for whole-brain counts and densities of markers of interest. In addition, original left and right hemisphere data is also returned in order for quality control analysis, or just in case the research is focused on a lateralized effect in the brain.