Regions Table Function
This a script for a function that takes raw data of multiple brain regions and cellular densities of a given population of cells, and returns a dataframe of brain regions in descending order of density target expression. It also reports the mean density for each region, sample size, standard deviation, and the standard error of the mean. This way, a researcher has a very good at-a-glance dataframe to see where a given class of cells is particularly enriched in the brain as well as the variability of the data.